+++ js/form.js
@@ -1,11 +1,17 @@
 import {isEscapeEvent} from './util.js';
 import {onIncreaseScaleClick, onDecreaseScaleClick, setDefaultScale} from './scale.js';
 import {onEffectChange, setDefaultEffect} from './effects.js';
-import {sendRequest} from './requests.js';
+import {sendNewPost} from './requests.js';
+import {showConnectionError, showSuccessMessage} from './message.js';
 
-const SEND_DATA_URL = 'https://28.javascript.pages.academy/kekstagram';
+// const SEND_DATA_URL = 'https://28.javascript.pages.academy/kekstagram';
 const MAX_SYMBOLS = 20;
 const MAX_HASHTAGS = 5;
+
+const SubmitButtonText = {
+  IDLE: 'Опубликовать',
+  SENDING: 'Опубликовываю...'
+};
 
 const pageBody = document.querySelector('body');
 const form = document.querySelector('.img-upload__form');
@@ -21,10 +27,6 @@ const submitButton = formUpload.querySelector('#upload-submit');
 const decreaseScaleElement = imageEditorDialog.querySelector('.scale__control--smaller');
 const increaseScaleElement = imageEditorDialog.querySelector('.scale__control--bigger');
 const effectsElement = imageEditorDialog.querySelector('.effects');
-
-let connectionErrorShown = false;
-let sendDataSucessShown = false;
-let formData = undefined;
 
 const pristine = new Pristine(formUpload, {
   classTo: 'img-upload__field-wrapper',
@@ -96,127 +98,14 @@ const onHashtagInput = () => {
   }
 };
 
-const onFormSubmit = (evt) => {
-  evt.preventDefault();
-  submitButton.disabled = true;
-  formData = new FormData(evt.target);
-  saveNewPost();
-};
-
 const isTextFieldInFocus = () => document.activeElement === hashTags || document.activeElement === comment;
 
-const onMouseClick = (evt) => {
-  if (connectionErrorShown) {
-
-    const errorElemenet = document.querySelector('.error__inner');
-    if (!evt.composedPath().includes(errorElemenet)) {
-
-      closeErrorBlock();
-    }
-  } else if (sendDataSucessShown) {
-    const successElemenet = document.querySelector('.success__inner');
-
-    if (!evt.composedPath().includes(successElemenet)) {
-
-      closeSuccessBlock();
-      closeImageEditor();
-    }
-  }
-};
-
-
 const onDocumentKeydown = (evt) => {
-  if (!isEscapeEvent(evt)) {
-    return;
-  }
 
-  if (connectionErrorShown) {
-
-    closeErrorBlock();
-  } else if (sendDataSucessShown) {
-
-    closeSuccessBlock();
-    closeImageEditor();
-  } else if (!isTextFieldInFocus()) {
-
+  if (isEscapeEvent(evt) && !isTextFieldInFocus()) {
     closeImageEditor();
   }
 };
-
-function closeErrorBlock() {
-  const errorElemenet = document.querySelector('.error');
-  pageBody.removeChild(errorElemenet);
-  document.removeEventListener('click', onMouseClick);
-  connectionErrorShown = false;
-}
-
-function closeSuccessBlock() {
-  const successElemenet = document.querySelector('.success');
-  pageBody.removeChild(successElemenet);
-  document.removeEventListener('click', onMouseClick);
-  sendDataSucessShown = false;
-}
-
-const changeTryAgainButtonAccessibility = (newValue) => {
-  const tryAgainButton = document.querySelector('.error__button');
-  tryAgainButton.disabled = newValue;
-};
-
-const onTryAgainButtonClick = () => {
-  changeTryAgainButtonAccessibility(true);
-  saveNewPost();
-};
-
-const onOkButtonClick = () => {
-  closeSuccessBlock();
-  closeImageEditor();
-};
-
-const showConnectionError = () => {
-  if (connectionErrorShown) {
-
-    changeTryAgainButtonAccessibility(false);
-    return;
-  }
-  connectionErrorShown = true;
-
-  submitButton.disabled = false;
-
-  const errorTemplate = document.querySelector('#error')
-    .content
-    .querySelector('.error');
-
-  const errorElement = errorTemplate.cloneNode(true);
-  pageBody.appendChild(errorElement);
-
-  const tryAgainButton = errorElement.querySelector('.error__button');
-  tryAgainButton.addEventListener('click', onTryAgainButtonClick);
-  document.addEventListener('click', onMouseClick);
-};
-
-const showSuccessMessage = () => {
-  sendDataSucessShown = true;
-
-  const sucсessTemplate = document.querySelector('#success')
-    .content
-    .querySelector('.success');
-
-  const sucсessElement = sucсessTemplate.cloneNode(true);
-  pageBody.appendChild(sucсessElement);
-
-  const okButton = sucсessElement.querySelector('.success__button');
-  okButton.addEventListener('click', onOkButtonClick);
-  document.addEventListener('click', onMouseClick);
-};
-
-function saveNewPost() {
-  const options = {
-    method: 'POST',
-    body: formData,
-  };
-  sendRequest(SEND_DATA_URL, showSuccessMessage, showConnectionError, options);
-  closeImageEditor();
-}
 
 function closeImageEditor () {
 
@@ -225,6 +114,7 @@ function closeImageEditor () {
 
   closeButton.removeEventListener('click', closeImageEditor);
   document.removeEventListener('keydown', onDocumentKeydown);
+
   form.removeEventListener('submit', onFormSubmit);
 
   increaseScaleElement.removeEventListener('click', onIncreaseScaleClick);
@@ -238,7 +128,7 @@ function closeImageEditor () {
   form.reset();
 }
 
-const openImageEditor = () => {
+const openImageEditor = () => {//открытие формы
 
   imageEditorDialog.classList.remove('hidden');
   pageBody.classList.add('modal-open');
@@ -249,10 +139,30 @@ const openImageEditor = () => {
   increaseScaleElement.addEventListener('click', onIncreaseScaleClick);
   decreaseScaleElement.addEventListener('click', onDecreaseScaleClick);
   effectsElement.addEventListener('change', onEffectChange);
+  form.addEventListener('submit', onFormSubmit);//вопросик имеет ли это место быть теперь?
+};
 
-  form.addEventListener('submit', onFormSubmit);
+const blockSubmitButton = () => {
+  submitButton.disabled = true;
+  submitButton.textContent = SubmitButtonText.SENDING;
+};
+
+const unblockSubmitButton = () => {
+  submitButton.disabled = false;
+  submitButton.textContent = SubmitButtonText.IDLE;
+};
+
+const saveNewPost = (evt) => sendNewPost(new FormData(evt.target), showSuccessMessage, showConnectionError);
+
+const onFormSubmit = (evt) => {
+  evt.preventDefault();
+  blockSubmitButton();
+  sendNewPost(new FormData(evt.target), showSuccessMessage, showConnectionError);
+  unblockSubmitButton();
 };
 
 inputHashtag.addEventListener('input', onHashtagInput);
 pristine.addValidator(inputHashtag, hashtagsHandler, getError, 2, false);
 uploadFile.addEventListener('change', openImageEditor);
+
+export {saveNewPost, closeImageEditor};
