+++ js/main.js
@@ -1,4 +1,7 @@
-import {loadPosts} from './render-thumbnail.js';
+import {renderThumbnails} from './render-thumbnail.js';
 import './form.js';
+import {loadPosts} from './requests.js';
+import { showConnectionErrorMessage} from './message.js';
 
-loadPosts();
+
+loadPosts(renderThumbnails, showConnectionErrorMessage);
