+++ js/render-thumbnail.js
@@ -1,8 +1,4 @@
 import {onPictureClick} from './big-picture.js';
-import {sendRequest} from './requests.js';
-
-const SOURCE_DATA_URL = 'https://28.javascript.pages.academy/kekstagram/data';
-const ERROR_SHOW_TIME = 7000;
 
 const container = document.querySelector('.pictures');
 const thumbnailTemplate = document.querySelector('#picture').content.querySelector('.picture');
@@ -40,27 +36,4 @@ const renderThumbnails = (thumbnails) => {
   container.appendChild(thumbnailFragment);
 };
 
-const showConnectionError = (errorDescription) => {
-  const errorContainer = document.createElement('div');
-  errorContainer.style.zIndex = '100';
-  errorContainer.style.position = 'absolute';
-  errorContainer.style.left = '0';
-  errorContainer.style.top = '0';
-  errorContainer.style.right = '0';
-  errorContainer.style.padding = '10px 3px';
-  errorContainer.style.fontSize = '30px';
-  errorContainer.style.textAlign = 'center';
-  errorContainer.style.backgroundColor = 'white';
-  errorContainer.style.color = 'red';
-  errorContainer.textContent = `При получении данных произошла ошибка ${errorDescription}`;
-
-  document.body.append(errorContainer);
-
-  setTimeout(() => {
-    errorContainer.remove();
-  }, ERROR_SHOW_TIME);
-};
-
-const loadPosts = () => sendRequest(SOURCE_DATA_URL, renderThumbnails, showConnectionError);
-
-export {loadPosts};
+export {renderThumbnails};
